% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocar_eucal.R
\name{biocar_eucal}
\alias{biocar_eucal}
\alias{biocar_eucal_cra}
\title{Desenvolvimento de Mudas de Eucalipto em Resposta \enc{à}{a}
    Doses de P e Biocarv\enc{ã}{a}o}
\format{
\code{biocar_eucal} é um \code{data.frame} com 90 observações
    e 20 variáveis, em que

\describe{

\item{\code{biocar}}{Fator de níveis categóricos que são as formas de
    fornecimento de P (fósforo) combinadas com biocarvão: S_BP - solo
    com biocarvão enriquecido com P (sem fósforo da adubação); SP_B -
    solo com biocarvão não enriquecido (sem fósforo no biocarvão) mas
    com P adicionado ao solo na forma de adubação; SP - fósforo
    fornecido ao solo como adubação sem adição de biocarvão.}

\item{\code{dosep}}{Fator de níveis métricos que é a dose de P
    (fósforo) aplicado ao solo na forma de adubação, em mg
    dm\eqn{^{-3}}. As doses não nulas seguem uma progressão
    geométrica: \code{1.5 * 2^(0:8)}.}

\item{\code{bloc}}{Fator categórico para controle local no
    experimento. Foram controladas as variações de ambiente da casa
    de vegetação, como exposição ao sol, e mão de obra ao longo do
    experimento, como irrigação e capina dos vasos e controle de
    doenças.}

\item{\code{alt}}{Altura (cm) das mudas.}

\item{\code{diam}}{Diâmetro (cm) das mudas.}

\item{\code{mfpa}}{Massa fresca de parte aérea (g).}

\item{\code{mfr}}{Massa fresca de raízes (g).}

\item{\code{mspa}}{Massa seca de parte aérea (g).}

\item{\code{msr}}{Massa seca de raízes (g).}

\item{\code{ph}}{pH em água.}

\item{\code{phkcl}}{pH em cloreto de potássio.}

\item{\code{pcz}}{Ponto de carga zero.}

\item{\code{p}}{Conteúdo de fósforo do solo, mg dm\eqn{^{-3}}.}

\item{\code{k}}{Conteúdo de potássio do solo, mg dm\eqn{^{-3}}.}

\item{\code{ca}}{Conteúdo de cálcio do solo, cmol\eqn{_{c}}
    dm\eqn{^{-3}}.}

\item{\code{mg}}{Conteúdo de magnésio do solo, cmol\eqn{_{c}}
    dm\eqn{^{-3}}.}

\item{\code{hal}}{Conteúdo de alumínio trocável (Al\eqn{^{3+}}),
    cmol\eqn{_{c}} dm\eqn{^{-3}}.}

\item{\code{are}}{Teor de areia do solo, g kg\eqn{^{-1}}.}

\item{\code{sil}}{Teor de silte do solo, g kg\eqn{^{-1}}.}

\item{\code{arg}}{Teor de argila do solo, g kg\eqn{^{-1}}.}

}

\code{biocar_eucal_cra} é um \code{data.frame} com 708 observações e
     5 variáveis, em que

\describe{

\item{\code{biocar}}{São os níveis de biocarvão, conforme já descrito
    acima. No entanto, as curva de retenção de água do solo só foram
    determinadas para os níveis \code{SP} e \code{SP_B}.}

\item{\code{dosep}}{Doses de fósforo, conforme já descrito acima.}

\item{\code{bloc}}{Níveis do fator bloco, conforme já descrito
    acima.}

\item{\code{tens}}{Tensão aplicada ao solo, kPa, para obtenção da
    correspondente umidade de equilíbrio.}

\item{\code{umid}}{Umidade do solo, m\eqn{^{3}} m\eqn{^{-3}} no
    equilíbrio com a tensão matricial aplicada.}

}
}
\source{
Milson Evaldo Serafim (orientador,
    \email{milson.serafim@cas.ifmt.edu.br}), Janaina Carvalho de
    Souza (Trabalho de Conclusão de Curso em Engenharia Florestal,
    2016).
}
\description{
Experimento foi instalado em casa de vegetação no mês de
    setembro do ano de 2015, no delineamento de blocos casualizados
    completos, utilizando como unidade experimental vasos de
    polietileno com 5 dm\eqn{^3} de solo com uma planta de
    eucalipto. A umidade do solo nos vasos foi mantida ente 65-55\%
    da capacidade de campo. O experimento teve duração de 90 dias,
    portanto, as variáveis agronômicas das mudas de eucalipto
    (dimensões e pesos) e as variáveis de solo (físicas, químicas e
    hídricas) foram determinadas aos 90 dias após a instalação do
    experimento.

    No experimento foram estudados dois fatores (fatorial 3
    \eqn{\times} 10): 3 formas de fornecimento de P (fósforo)
    combinadas com 10 doses de P para o desenvolvimento mudas de
    eucalipto.

    As formas de fornecimento de P (fator categórico) foram três (3):
    1) por meio da adubação (\code{SP}, fósforo aplicado ao solo), 2)
    da aplicação de biocarvão enriquecido de P (\code{S_BP}, o
    fósforo está no biocarvão) e 3) da aplicação de biocarvão não
    enriquecido com o fósforo adicionado por adubação (\code{SP_B}).

    As doses não nulas de P (fator métrico) são elementos de uma
    progressão geométrica: \eqn{dose(i) = 1,5 \times 2^{i}, i = 0, 1,
    \ldots, 8}, totalizando, com a dose nula (P = 0), 10 níveis.

    Amostras de solo indeformadas foram extraídas dos vasos para
    determinação da curva de retenção de água (CRA) do solo. Depois
    de saturadas, as amostras foram pesadas e submetidas aos
    seguintes potenciais: -1, -2, -4, -6, -10 kPa, em funis de placa
    porosa na unidade de sucção, e -33, -66, -300 e -1500 kPa, no
    aparelho extratores de Richards, segundo Embrapa (1997). Após
    atingir o equilíbrio hídrico em cada potencial, as amostras foram
    pesadas e submetidas ao potencial subsequente, constituindo
    método por secamento. Após o último potencial, as amostras foram
    secas em estufa a \eqn{\pm} 105 \eqn{^\circ}C, por 24 horas, para
    determinação do conteúdo de água residual do solo (umidade
    residual da última tensão).
}
\examples{

str(biocar_eucal)
str(biocar_eucal_cra)

}
\references{
PAULA, J. L. DE; DUARTE, M. N. Manual de métodos de
    análise de solo. 2nd ed. Rio de Janeiro: EMBRAPA-CNPS, 1997.
}
