% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafe_pedotrans.R
\name{cafe_pedotrans}
\alias{cafe_pedotrans}
\title{Modelagem da capacidade de suporte de carga de latossolos com
    diferente mineralogia}
\format{A \code{data.frame} with 180 observations and 12 variables,
    in which

\describe{

\item{\code{solo}}{Fator experimental qualitivo que indica o tipo de
    solo.}

\item{\code{tensao}}{Tensão matricial aplicada nas amostras de solo:
    6, 33, 100, 500, 1500, 3000 kPa.}

\item{\code{posi}}{Fator experimental qualitativo que é a posição em
    relação as linhas de cultivo do café: linha e entrelinha.}

\item{\code{prof}}{0-5,10-15,20-25m}

\item{\code{rep}}{Variável numérica que distingue entre as repetições
    do mesmo ponto experimental.}

\item{\code{dsi}}{Variável medida nas amostras de solo que é a
    densidade do solo inicial, em g cm\eqn{^{-3}}.}

\item{\code{dsp}}{Variável medida nas amostras de solo que é a
    densidade na pressão de preconsolidação, em g cm\eqn{^{-3}}.}

\item{\code{thetai}}{Variável medida nas amostras de solo que é o
    conteúdo de água, cm\eqn{^3} cm\eqn{^{-3}}.}

\item{\code{ppc}}{Valor de pressão de preconsolidação observado na
    amostra de solo, em kPa.}

\item{\code{trt}}{Variável criada pela combinação dos níveis de solo,
    posição e profundidade para indicar cada ponto experimental ou
    tratamento presente no experimento.}

\item{\code{ue}}{Variável criada pela combinação dos níveis de solo,
    posição, profundidade e repetição para indicar cada curva
    presente no experimento.}

}}
\source{
Carla Eloize Carducci, \email{elocarducci@gmail.com},
    \url{http://buscatextual.cnpq.br/buscatextual/visualizacv.do?id=K4713341P8};
    Geraldo Cesar de Oliveria, \email{geraldooliveira@dcs.ufla.br},
    \url{http://buscatextual.cnpq.br/buscatextual/visualizacv.do?id=K4728903U6};
    Eduardo Costa Severiano,
    \email{eduardo.severiano@ifgoiano.edu.br},
    \url{http://buscatextual.cnpq.br/buscatextual/visualizacv.do?id=K4777485Y2}.
}
\description{
O experimento foi realizado em São Roque de Minas, Minas
    Gerais (20 15' 43" S, 46 22' 17" E) no delineamento experimental
    inteiramente casualizado com 3 repetições. Foram avaliados o
    efeito de tipo de solo, dois experimentos independentes com a
    mesma estrutura foram instalados em propriedades com solo
    diferente. Foi avaliada a posição de amostragem com níveis linha
    do cultivo do café onde é aplicado alta dose de gesso e
    entrelinha do cultivo do café que tem manejo feito com
    gramíneas. Também avaliou-se a profundidade do solo com amostras
    coletadas em 3 camadas. As variáveis resposta estão descritas a
    seguir.

    As hipóteses que motivaram essa investigação são i) determinar
    ser o sistema de manejo alterou a capacidade de suporte de carga
    dos solos e ii) se o sistema de manejo foi capaz de padronizar a
    estrutura de solos de mineralogia distinta.
}
\examples{

library(lattice)

xyplot(ppc ~ tensao | trt,
       groups = rep,
       data = cafe_pedotrans,
       type = "o",
       scales = list(x = list(log = TRUE)),
       layout = c(3, NA))

xyplot(ppc ~ tensao | solo + prof,
       groups = posi,
       data = cafe_pedotrans,
       type = c("p", "a"),
       scales = list(x = list(log = 10)),
       auto.key = TRUE)

}
