% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EACS.R
\docType{package}
\name{EACS}
\alias{EACS}
\alias{EACS-package}
\title{Estatística Aplicada à Ciência do Solo}
\description{
Pacote que documenta conjuntos de dados de Projetos de
    Pesquisa na Ciência do Solo. O pacote possui as análises dos
    dados no formato de vinheta com o objetivo de disponilizar essas
    análises como molde ou inspiração para problemas que sejam
    semelhantes.
}
\examples{
\dontrun{

# Carrega o pacote.
library(EACS)

# Lista os objetos do pacote.
ls("package:EACS")

# Abre a documentação do pacote.
help(package = "EACS", help_type = "html")

# Abre a lista de vinhetas no navegador.
browseVignettes(package = "EACS")

# Exibe a lista no console.
vignette(package = "EACS")

# Abre uma das vinhetas.
vignette(topic = "v01_poisson", package = "EACS")

# Citação do pacote.
citation("EACS")

}
}

