% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fat2_adic.R
\name{fat2_adic}
\alias{fat2_adic}
\title{Experimento fatorial duplo com um tratamento adicional como
    testemunha}
\format{
Um \code{data.frame} com 36 observações e 4 variáveis, em que

\describe{

\item{\code{fonte}}{Fator categórico que representa os as fontes
    utilizadas de adubação. O controle é o último nível do fator e
    representa as parcelas que não receberam adubação.}

\item{\code{modo}}{Forma de aplicação das fontes de adubação. Esse
    fator também tem o controle como último nível e representa as
    parcelas que não tiveram adubação, pontanto, o fator modo sequer
    se aplica.}

\item{\code{bloc}}{Fator categórico que representa os blocos do
    experimento.}

\item{\code{prod}}{Variável resposta do experimento, produção da
    cultura. Informações não divulgadas}

}
}
\source{
Dados cedidos por Fábio B. Ono
    (\email{fabioono@fundacaomt.com.br}) para propósitos
    didáticos. O uso dos dados para outros fins só deverá ser feito
    após aprovação.
}
\description{
Este experimento corresponde a um fatorial duplo 4
    \eqn{\times} 2 com um controle negativo. No fatorial são
    estudados 4 fontes de adubação cruzados com dois modos de
    aplicação: no sulco e a lanço. O nível controle corresponde à uma
    testemunha negativa uma vez que ce antecipa uma baixa produção
    devido a não realização da adubação. O experimento foi realizado
    do delineamento de blocos casualizados completos com 4
    repetições.
}
\examples{

xtabs(~fonte + modo, data = fat2_adic)

library(lattice)
library(latticeExtra)

resizePanels(
    xyplot(prod ~ fonte | modo,
           data = fat2_adic,
           type = c("p", "a"),
           layout = c(NA, 1),
           scales = list(x = list(relation = "free"))),
    w = c(4, 4, 1))

resizePanels(
    xyplot(prod ~ modo | fonte,
           data = fat2_adic,
           type = c("p", "a"),
           layout = c(NA, 1),
           scales = list(x = list(relation = "free"))),
    w = c(2, 2, 2, 2, 1))

}
