% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teca_coords.R
\name{teca_coords}
\alias{teca_coords}
\title{Coordenadas dos s\enc{í}{i}tios de Teca}
\format{Um \code{data.frame} com 50 observações e 5 variáveis, em que

\describe{

\item{\code{loc}}{Variável identificadora do ponto amostral.}

\item{\code{longitude}}{Longitude em UTM.}

\item{\code{latitude}}{Latitude em UTM.}

\item{\code{vol}}{Volume de madeira em metros cúbicos.}

\item{\code{ima}}{Incremento médio anual, refere-se ao ganho em
    diâmetro da arvore. Embora tenha relação com o volume total, o
    grau da relação depende da maior ou menor homogeneidade de altura
    das árvores do sítio.}

}

Para mais informações, consulte a documentação das demais tabelas:
\code{\link{teca_arv}}, \code{\link{teca_ndvi}},
\code{\link{teca_qui}}, \code{\link{teca_cra}},
\code{\link{teca_crapar}}, \code{\link{teca_gran}}.}
\source{
Milson Evaldo Serafim.
}
\description{
Tabela que contém as coordenada (UTM) dos sítios de
    Teca e variáveis de produção.
}
\examples{

library(lattice)

xyplot(vol ~ ima, data = teca_coords)

xyplot(latitude ~ longitude,
       data = teca_coords,
       cex = 4 * teca_coords$vol,
       scales = list(x = list(rot = 90)))

xyplot(latitude ~ longitude,
       data = teca_coords,
       cex = 0.1 * teca_coords$ima,
       scales = list(x = list(rot = 90)))

}
